refactor: move database creation code around
Now all the code that is required to create a new database lives under
`server::database`, so it can easily be used for tests that don't
involve a server.
